Fix PlanNode.Plan.Actual Rows int64 unmarshal error on fractional ave… (#19)
* Fix PlanNode.Plan.Actual Rows int64 unmarshal error on fractional averages

PostgreSQL computes "Actual Rows" as total_rows / nloops and always
formats it with 2 decimal digits in JSON EXPLAIN output (ExplainNode()
in explain.c), for every node type. When Actual Loops > 1 and the
division doesn't come out even (e.g. 3079.00, or 242.31 for
high-loop-count inner scans), the value is a genuine non-integer float.
pgplan modeled ActualRows as int64, so json.Unmarshal rejected any plan
with loops > 1, which is common for nested loop inner sides — this
crashed `pgplan compare`/`analyze` on real-world plans.

Change ActualRows (and, defensively, RowsRemovedByFilter/
RowsRemovedByJoinFilter, which share the same instrumentation path)
to float64, and propagate the type through the analyzer rules,
comparator diff/result structs, and text output formatting. Other
int64 fields (Plan Rows, Actual Loops, buffer/block counts, hash/sort
stats) were audited against PostgreSQL's source and confirmed to
always be emitted as whole numbers, so they're left unchanged.

Adds regression tests covering fractional values mixed with plain
integers in the same plan, and confirms the fix holds across multiple
node types (Seq Scan, Index Scan, Index Only Scan, Bitmap Heap Scan,
Hash Join, Sort, Aggregate) since the underlying Postgres code path is
shared by all node types.


* Revert RowsRemovedByFilter/RowsRemovedByJoinFilter to int64

Both go through show_instrumentation_count() in PostgreSQL's explain.c,
which divides nfiltered by nloops just like Actual Rows but formats the
result with 0 decimal digits (ExplainPropertyFloat(..., 0, es)) instead
of 2. PostgreSQL therefore never emits a fractional value for these two
fields, so the earlier float64 change was unnecessary defensive slack —
int64 is correct and matches actual PostgreSQL output.


---------

+++ internal/analyzer/context.go
@@ -16,7 +16,7 @@ type CTEInfo struct {
 	Name          string
 	Node          *plan.PlanNode
 	EstimatedRows int64
-	ActualRows    int64
+	ActualRows    float64
 }
 
 type NodeRef struct {
